+++ crates/cubed-application/src/ports/mod.rs
@@ -1,5 +1,7 @@
 pub mod file_system;
+pub mod java_manager;
 pub mod server_repository;
 
 pub use file_system::FileSystemManager;
+pub use java_manager::{JavaInstallation, JavaManager};
 pub use server_repository::ServerRepository;