+++ crates/cubed-application/src/use_cases/mod.rs
@@ -2,6 +2,7 @@ pub mod create_server;
 pub mod delete_server;
 pub mod init_file_system;
 pub mod restart_server;
+pub mod select_java;
 pub mod start_server;
 pub mod stop_server;
 
@@ -9,5 +10,6 @@ pub use create_server::{CreateServer, CreateServerInput};
 pub use delete_server::DeleteServer;
 pub use init_file_system::InitFileSystem;
 pub use restart_server::RestartServer;
+pub use select_java::SelectJava;
 pub use start_server::StartServer;
 pub use stop_server::StopServer;